# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TalentRank API
  version: 0.1.0
  description: Human-in-the-loop candidate ranking API. All tenant resources are workspace scoped.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      responses:
        "200": { description: Workspace jobs, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/JobListItem" } } } } } } }
    post:
      tags: [Jobs]
      operationId: createJob
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/JobWrite" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "422": { $ref: "#/components/responses/ValidationError" }
  /jobs/{jobId}:
    parameters: [{ $ref: "#/components/parameters/JobId" }]
    get:
      tags: [Jobs]
      operationId: getJob
      responses:
        "200": { description: Job, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/requirements:
    parameters: [{ $ref: "#/components/parameters/JobId" }]
    put:
      tags: [Jobs]
      operationId: replaceRequirements
      requestBody: { required: true, content: { application/json: { schema: { type: array, maxItems: 200, items: { $ref: "#/components/schemas/Requirement" } } } } }
      responses:
        "200": { description: New immutable job version, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: talentrank_refresh }
  parameters:
...
    JobId: { in: path, name: jobId, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    NotFound: { description: Resource not found in workspace, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    ValidationError: { description: Invalid request, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        errors: { type: array, items: { type: object, additionalProperties: true } }
...
    JobStatus: { type: string, enum: [draft, active, paused, closed, archived] }
    JobWrite:
      type: object
      required: [title, description]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, minLength: 20, maxLength: 100000 }
        location: { type: string, maxLength: 200 }
        employment_type: { type: string, enum: [full_time, part_time, contract, internship] }
    JobListItem:
      type: object
      required: [id, title, location, employment_type, status, candidate_count, created_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        location: { type: string }
        employment_type: { type: string, enum: [full_time, part_time, contract, internship] }
        status: { $ref: "#/components/schemas/JobStatus" }
        candidate_count: { type: integer, minimum: 0 }
        last_ranking_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
    Job:
      allOf:
        - { $ref: "#/components/schemas/JobListItem" }
        - type: object
          required: [description, current_version, requirements, updated_at]
          properties:
            description: { type: string }
            summary: { type: string }
            seniority: { type: [string, "null"] }
            current_version: { type: integer, minimum: 1 }
            requirements: { type: array, items: { $ref: "#/components/schemas/Requirement" } }
            updated_at: { type: string, format: date-time }
    Requirement:
      type: object
      required: [requirement_key, name, requirement_type, priority, weight, is_hard_constraint, description]
      properties:
        id: { type: string, format: uuid }
        requirement_key: { type: string, maxLength: 100 }
        name: { type: string, minLength: 1, maxLength: 200 }
        requirement_type: { type: string, enum: [skill, experience, education, certification, language, location, other] }
        priority: { type: string, enum: [must_have, nice_to_have] }
        weight: { type: number, minimum: 0, maximum: 10 }
        minimum_years: { type: number, minimum: 0, maximum: 80 }
        is_hard_constraint: { type: boolean }
        description: { type: string, maxLength: 2000 }
...